Minor refactorings
